ai_news_tracker/backend/requirements.txt
# AI News Tracker - Backend Dependencies

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0
chromadb==0.4.18

# Web Scraping
playwright==1.40.0
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.3
openpyxl==3.1.2

# AI Services
openai==1.6.1
anthropic==0.7.8

# Task Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
